# docs/openapi.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "HistoryTalk API",
    "version": "1.0.0",
    "description": "Vietnamese history educational AI platform — backend REST API"
  },
  "servers": [
    {
      "url": "http://localhost:5000/api/v1",
      "description": "Local development"
    },
    {
      "url": "https://api.historytalk.dev/api/v1",
      "description": "Production"
    }
  ],
  "components": {
    "securitySchemes": {
      "BearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      }
    },
    "schemas": {
      "ErrorResponse": {
        "type": "object",
        "properties": {
          "success": {
            "type": "boolean",
            "example": false
          },
          "message": {
            "type": "string",
            "example": "Something went wrong"
          },
          "data": {
            "nullable": true,
            "example": null
          },
          "timestamp": {
            "type": "string",
            "example": "2025-05-22T10:00:00.000Z"
          }
        }
      },
...
      "UserProfile": {
        "type": "object",
        "properties": {
          "_id": {
            "type": "string",
            "example": "64a1b2c3d4e5f6789abcdef0"
          },
          "userName": {
            "type": "string",
            "example": "Nguyen Van A"
          },
          "email": {
            "type": "string",
            "format": "email",
            "example": "user@example.com"
          },
          "role": {
            "type": "string",
            "enum": [
              "CUSTOMER",
              "CONTENT_ADMIN",
              "SYSTEM_ADMIN"
            ],
            "example": "CUSTOMER"
          },
          "token": {
            "type": "integer",
            "example": 10
          },
          "tierId": {
            "type": "string",
            "example": "64a1b2c3d4e5f6789abcdef1"
          },
          "lastActiveDate": {
            "type": "string",
            "format": "date-time"
          },
          "createdAt": {
            "type": "string",
            "format": "date-time"
          },
          "updatedAt": {
            "type": "string",
            "format": "date-time"
          }
        }
      },
      "UpdateProfileBody": {
        "type": "object",
        "properties": {
          "userName": {
            "type": "string",
            "minLength": 2,
            "example": "Updated Name"
          }
        }
      }
    }
  },
  "paths": {
...
    "/system-admin/dashboard/users": {
      "get": {
        "tags": [
          "System Admin Dashboard"
        ],
        "summary": "Get user analytics",
        "security": [
          {
            "BearerAuth": []
          }
        ]
      }
    },
...
    "/quizzes/{quizId}": {
      "get": {
        "tags": [
          "Quizzes"
        ],
        "summary": "Get quiz by ID",
        "parameters": [
          {
            "in": "path",
            "name": "quizId",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ]
      }
    },
...
    "/staff/quizzes": {
      "get": {
        "tags": [
          "Staff Quizzes"
        ],
        "summary": "List quizzes for staff",
        "security": [
          {
            "BearerAuth": []
          }
        ]
      },
      "post": {
        "tags": [
          "Staff Quizzes"
        ],
        "summary": "Create a quiz",
        "security": [
          {
            "BearerAuth": []
          }
        ]
      }
    },
    "/staff/quizzes/{quizId}": {
      "get": {
        "tags": [
          "Staff Quizzes"
        ],
        "summary": "Get quiz detail with all questions",
        "security": [
          {
            "BearerAuth": []
          }
        ],
        "parameters": [
          {
            "in": "path",
            "name": "quizId",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ]
      },
      "put": {
        "tags": [
          "Staff Quizzes"
        ],
        "summary": "Update quiz metadata",
        "security": [
          {
            "BearerAuth": []
          }
        ],
        "parameters": [
          {
            "in": "path",
            "name": "quizId",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ]
      },
      "delete": {
        "tags": [
          "Staff Quizzes"
        ],
        "summary": "Delete a quiz",
        "security": [
          {
            "BearerAuth": []
          }
        ],
        "parameters": [
          {
            "in": "path",
            "name": "quizId",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ]
      }
    },
    "/staff/quizzes/{quizId}/soft-delete": {
      "patch": {
        "tags": [
          "Staff Quizzes"
        ],
        "summary": "Soft delete a quiz",
        "security": [
          {
            "BearerAuth": []
          }
        ],
        "parameters": [
          {
            "in": "path",
            "name": "quizId",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ]
      }
    },
    "/staff/quizzes/{quizId}/restore": {
      "patch": {
        "tags": [
          "Staff Quizzes"
        ],
        "summary": "Restore a soft-deleted quiz",
        "security": [
          {
            "BearerAuth": []
          }
        ],
        "parameters": [
          {
            "in": "path",
            "name": "quizId",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ]
      }
    },
    "/staff/quizzes/{quizId}/questions": {
      "post": {
        "tags": [
          "Staff Quizzes"
        ],
        "summary": "Add question to quiz",
        "security": [
          {
            "BearerAuth": []
          }
        ],
        "parameters": [
          {
            "in": "path",
            "name": "quizId",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ]
      }
    },
    "/staff/quizzes/{quizId}/questions/{questionId}": {
      "put": {
        "tags": [
          "Staff Quizzes"
        ],
        "summary": "Update a question",
        "security": [
          {
            "BearerAuth": []
          }
        ],
        "parameters": [
          {
            "in": "path",
            "name": "quizId",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "in": "path",
            "name": "questionId",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ]
      },
      "delete": {
        "tags": [
          "Staff Quizzes"
        ],
        "summary": "Delete a question",
        "security": [
          {
            "BearerAuth": []
          }
        ],
        "parameters": [
          {
            "in": "path",
            "name": "quizId",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "in": "path",
            "name": "questionId",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ]
      }
    },
    "/users/me": {
      "get": {
        "tags": [
          "Users"
        ],
        "summary": "Get the authenticated user's profile",
        "security": [
          {
            "BearerAuth": []
          }
        ],
        "responses": {
          "200": {
            "description": "Profile fetched",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "success": {
                      "type": "boolean",
                      "example": true
                    },
                    "message": {
                      "type": "string"
                    },
                    "data": {
                      "type": "object",
                      "properties": {
                        "user": {
                          "$ref": "#/components/schemas/UserProfile"
                        }
                      }
                    },
                    "timestamp": {
                      "type": "string"
                    }
                  }
                }
              }
            }
          },
          "401": {
            "description": "Missing or invalid token",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      },
      "patch": {
        "tags": [
          "Users"
        ],
        "summary": "Update the authenticated user's profile",
        "security": [
          {
            "BearerAuth": []
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateProfileBody"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Profile updated",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "success": {
                      "type": "boolean",
                      "example": true
                    },
                    "message": {
                      "type": "string"
                    },
                    "data": {
                      "type": "object",
                      "properties": {
                        "user": {
                          "$ref": "#/components/schemas/UserProfile"
                        }
                      }
                    },
                    "timestamp": {
                      "type": "string"
                    }
                  }
                }
              }
            }
          },
          "401": {
            "description": "Missing or invalid token",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      }
    },
    "/users/me/password": {
      "patch": {
        "tags": [
          "Users"
        ],
        "summary": "Change the authenticated user's password",
        "security": [
          {
            "BearerAuth": []
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "currentPassword": {
                    "type": "string"
                  },
                  "newPassword": {
                    "type": "string"
                  },
                  "confirmPassword": {
                    "type": "string"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Password changed successfully"
          },
          "400": {
            "description": "Passwords do not match or incorrect current password"
          }
        }
      }
    },
    "/users": {
      "get": {
        "tags": [
          "Users"
        ],
        "summary": "List all users (SystemAdmin only)",
        "security": [
          {
            "BearerAuth": []
          }
        ]
      }
    },
    "/users/{id}": {
      "get": {
        "tags": [
          "Users"
        ],
        "summary": "Get a user by ID (SystemAdmin only)",
        "security": [
          {
            "BearerAuth": []
          }
        ],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ]
      },
      "patch": {
        "tags": [
          "Users"
        ],
        "summary": "Update a user (SystemAdmin only)",
        "security": [
          {
            "BearerAuth": []
          }
        ],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ]
      }
    },
    "/users/{id}/role": {
      "patch": {
        "tags": [
          "Users"
        ],
        "summary": "Update a user's role (SystemAdmin only)",
        "security": [
          {
            "BearerAuth": []
          }
        ],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ]
      }
    }
  },
...
}